+++ src/index.js
@@ -43,6 +43,7 @@ export {
   simulateBattle
 } from './battle-simulation.js';
 
+export * from './wallet-accounting.js';
 export * from './asset-gacha.js';
 export * from './client-view-model.js';
 export * from './client/index.js';